pyproject.toml
[project]
name = "mlx-lama"
version = "0.1.0"
description = "Ollama-compatible CLI for MLX models on Apple Silicon"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Tim Labsky" }
]
keywords = ["mlx", "llm", "apple-silicon", "ollama", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "huggingface-hub>=0.20.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
mlx = ["mlx-lm>=0.19.0"]
vllm = ["vllm-mlx>=0.1.0"]
dev = ["pytest>=7.0.0", "ruff>=0.1.0"]

[project.scripts]
mlx-lama = "mlx_lama.cli:app"

[project.urls]
Homepage = "https://github.com/timotejlabsky/mlx-lama"
Repository = "https://github.com/timotejlabsky/mlx-lama"
Issues = "https://github.com/timotejlabsky/mlx-lama/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlx_lama"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
